javademo/src/main/java/thread/lock/DelayQueueUsingTimedSemaphore.java
package thread.lock;

import java.util.concurrent.TimeUnit;
import org.apache.commons.lang3.concurrent.TimedSemaphore;

/**
 * https://www.baeldung.com/java-semaphore
 */
public class DelayQueueUsingTimedSemaphore {

    private TimedSemaphore semaphore;

    public DelayQueueUsingTimedSemaphore(long period, int slotLimit) {
        this.semaphore = new TimedSemaphore(period, TimeUnit.SECONDS, slotLimit);
    }

    public boolean tryAdd() {
        return semaphore.tryAcquire();
    }

    public int availableSlots() {
        return semaphore.getAvailablePermits();
    }
}